feat: `canBeChoseong`, `canBeJungseong`, `canBeJongseong` 을 utils에서 별도 함수로 분리합니다. (#193)
* canBe series를 별도 함수로 분리합니다

* add change set
