+++ .changeset/pretty-apes-destroy.md
@@ -1,0 +1,5 @@
+---
+"es-hangul": major
+---
+
+canBeChoseong, canBeJungseong, canBeJongseong 을 utils에서 별도 함수로 분리합니다.
