+++ src/index.ts
@@ -1,13 +1,14 @@
 export { assembleHangul } from './assemble';
+export { canBeChoseong, canBeJongseong, canBeJungseong } from './canBe';
 export { choseongIncludes } from './choseongIncludes';
 export { chosungIncludes } from './chosungIncludes';
 export { combineHangulCharacter, combineVowels, curriedCombineHangulCharacter } from './combineHangulCharacter';
 export { convertQwertyToHangul, convertQwertyToHangulAlphabet } from './convertQwertyToHangulAlphabet';
 export { disassemble, disassembleToGroups } from './disassemble';
 export { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
+export { getChoseong } from './getChoseong';
 export { josa } from './josa';
 export { removeLastHangulCharacter } from './removeLastHangulCharacter';
 export { romanize } from './romanize';
 export { standardizePronunciation } from './standardizePronunciation';
 export { susa } from './susa';
-export { getChoseong } from './getChoseong';
